refactor: re-organize core LLM files
Re-organize files, splitting the workflow-based entity recognition from the more agentic dependency parsing.

Also drop some dead code and associated tests.
