+++ test/test_models.py
@@ -1,221 +1,8 @@
-import json
-import pytest
-from recipe_board.agents.models import parse_recipe, parse_dependencies, ingredients_and_equipment_from_parsed_recipe
+from recipe_board.agents.parsing_agent import parse_dependencies
+from recipe_board.agents.entity_workflow import parse_recipe
 from recipe_board.core.state import RecipeSessionState, ParsingState
 from recipe_board.core.recipe import Ingredient, Equipment, BasicAction
 from smolagents.agent_types import AgentText
-
-
-class TestIngredientsAndEquipmentFromParsedRecipe:
-    """Test suite for ingredients_and_equipment_from_parsed_recipe function."""
-
-    def test_valid_complete_json(self):
-        """Test parsing valid JSON with complete data."""
-        json_data = json.dumps({
-            "ingredients": [
-                {"name": "flour", "amount": 2.0, "unit": "cups", "modifiers": ["all-purpose"]},
-                {"name": "salt", "amount": 1.0, "unit": "tsp", "modifiers": None}
-            ],
-            "equipment": [
-                {"name": "mixing bowl", "required": True, "modifiers": "large"},
-                {"name": "whisk", "required": False, "modifiers": None}
-            ]
-        })
-
-        ingredients, equipment = ingredients_and_equipment_from_parsed_recipe(json_data)
-
-        # Verify ingredients
-        assert len(ingredients) == 2
-        assert ingredients[0].name == "flour"
-        assert ingredients[0].amount == 2.0
-        assert ingredients[0].unit == "cups"
-        assert ingredients[0].modifiers == ["all-purpose"]
-        assert "2.0 cups flour all-purpose" in ingredients[0].raw_text
-
-        assert ingredients[1].name == "salt"
-        assert ingredients[1].amount == 1.0
-        assert ingredients[1].unit == "tsp"
-        assert ingredients[1].modifiers == []
-
-        # Verify equipment
-        assert len(equipment) == 2
-        assert equipment[0].name == "mixing bowl"
-        assert equipment[0].required == True
-        assert equipment[0].modifiers == "large"
-
-        assert equipment[1].name == "whisk"
-        assert equipment[1].required == False
-        assert equipment[1].modifiers == None
-
-    def test_string_modifiers_conversion(self):
-        """Test conversion of string modifiers to list."""
-        json_data = json.dumps({
-            "ingredients": [
-                {"name": "flour", "amount": 2.0, "unit": "cups", "modifiers": "all-purpose"}
-            ],
-            "equipment": []
-        })
-
-        ingredients, equipment = ingredients_and_equipment_from_parsed_recipe(json_data)
-
-        assert len(ingredients) == 1
-        assert ingredients[0].modifiers == ["all-purpose"]
-
-    def test_missing_optional_fields(self):
-        """Test handling of missing optional fields."""
-        json_data = json.dumps({
-            "ingredients": [
-                {"name": "salt"}  # Missing amount, unit, modifiers
-            ],
-            "equipment": [
-                {"name": "bowl"}  # Missing required, modifiers
-            ]
-        })
-
-        ingredients, equipment = ingredients_and_equipment_from_parsed_recipe(json_data)
-
-        assert len(ingredients) == 1
-        assert ingredients[0].name == "salt"
-        assert ingredients[0].amount == None
-        assert ingredients[0].unit == None
-        assert ingredients[0].modifiers == []
-        assert ingredients[0].raw_text == "salt"
-
-        assert len(equipment) == 1
-        assert equipment[0].name == "bowl"
-        assert equipment[0].required == True  # Default value
-        assert equipment[0].modifiers == None
-
-    def test_empty_arrays(self):
-        """Test handling of empty ingredients and equipment arrays."""
-        json_data = json.dumps({
-            "ingredients": [],
-            "equipment": []
-        })
-
-        ingredients, equipment = ingredients_and_equipment_from_parsed_recipe(json_data)
-
-        assert len(ingredients) == 0
-        assert len(equipment) == 0
-
-    def test_invalid_json(self):
-        """Test handling of invalid JSON."""
-        with pytest.raises(ValueError, match="Invalid JSON"):
-            ingredients_and_equipment_from_parsed_recipe("not valid json")
-
-    def test_missing_required_keys(self):
-        """Test handling of missing required top-level keys."""
-        # Missing equipment key
-        json_data = json.dumps({"ingredients": []})
-        with pytest.raises(ValueError, match="must contain 'ingredients' and 'equipment' keys"):
-            ingredients_and_equipment_from_parsed_recipe(json_data)
-
-        # Missing ingredients key
-        json_data = json.dumps({"equipment": []})
-        with pytest.raises(ValueError, match="must contain 'ingredients' and 'equipment' keys"):
-            ingredients_and_equipment_from_parsed_recipe(json_data)
-
-    def test_non_object_json(self):
-        """Test handling of non-object JSON."""
-        with pytest.raises(ValueError, match="JSON must be an object"):
-            ingredients_and_equipment_from_parsed_recipe("[]")
-
-        with pytest.raises(ValueError, match="JSON must be an object"):
-            ingredients_and_equipment_from_parsed_recipe("\"string\"")
-
-    def test_non_array_ingredients_equipment(self):
-        """Test handling of non-array ingredients/equipment."""
-        json_data = json.dumps({
-            "ingredients": "not an array",
-            "equipment": []
-        })
-        with pytest.raises(ValueError, match="'ingredients' must be an array"):
-            ingredients_and_equipment_from_parsed_recipe(json_data)
-
-        json_data = json.dumps({
-            "ingredients": [],
-            "equipment": "not an array"
-        })
-        with pytest.raises(ValueError, match="'equipment' must be an array"):
-            ingredients_and_equipment_from_parsed_recipe(json_data)
-
-    def test_invalid_ingredient_items_skipped(self):
-        """Test that invalid ingredient items are skipped with warnings."""
-        json_data = json.dumps({
-            "ingredients": [
-                {"name": "valid flour", "amount": 2.0},
-                "invalid string item",
-                {"name": "valid salt", "amount": 1.0}
-            ],
-            "equipment": []
-        })
-
-        ingredients, equipment = ingredients_and_equipment_from_parsed_recipe(json_data)
-
-        # Should only get the valid ingredients
-        assert len(ingredients) == 2
-        assert ingredients[0].name == "valid flour"
-        assert ingredients[1].name == "valid salt"
-
-    def test_invalid_equipment_items_skipped(self):
-        """Test that invalid equipment items are skipped with warnings."""
-        json_data = json.dumps({
-            "ingredients": [],
-            "equipment": [
-                {"name": "valid bowl"},
-                "invalid string item",
-                {"name": "valid spoon"}
-            ]
-        })
-
-        ingredients, equipment = ingredients_and_equipment_from_parsed_recipe(json_data)
-
-        # Should only get the valid equipment
-        assert len(equipment) == 2
-        assert equipment[0].name == "valid bowl"
-        assert equipment[1].name == "valid spoon"
-
-    def test_raw_text_generation(self):
-        """Test proper raw_text generation from components."""
-        json_data = json.dumps({
-            "ingredients": [
-                {"name": "flour", "amount": 2.5, "unit": "cups", "modifiers": ["all-purpose", "sifted"]},
-                {"name": "water", "amount": 1, "unit": "cup"},
-                {"name": "salt"}
-            ],
-            "equipment": []
-        })
-
-        ingredients, equipment = ingredients_and_equipment_from_parsed_recipe(json_data)
-
-        assert "2.5 cups flour all-purpose sifted" == ingredients[0].raw_text
-        assert "1 cup water" == ingredients[1].raw_text
-        assert "salt" == ingredients[2].raw_text
-
-    def test_unique_ids_generated(self):
-        """Test that unique IDs are generated for each object."""
-        json_data = json.dumps({
-            "ingredients": [
-                {"name": "flour", "amount": 2.0},
-                {"name": "salt", "amount": 1.0}
-            ],
-            "equipment": [
-                {"name": "bowl"},
-                {"name": "spoon"}
-            ]
-        })
-
-        ingredients, equipment = ingredients_and_equipment_from_parsed_recipe(json_data)
-
-        # All IDs should be unique
-        all_ids = [ing.id for ing in ingredients] + [eq.id for eq in equipment]
-        assert len(all_ids) == len(set(all_ids))  # No duplicates
-
-        # All IDs should be valid UUIDs (basic check)
-        for id_str in all_ids:
-            assert len(id_str) == 36  # UUID string length
-            assert id_str.count('-') == 4  # UUID format
-
 
 class TestParseRecipeEquipment:
     """Test suite for parse_recipe function (new state-based API)."""
@@ -241,7 +28,7 @@ class TestParseRecipeEquipment:
             return mock_response
 
         monkeypatch.setattr(
-            "recipe_board.agents.models.InferenceClient.text_generation",
+            "recipe_board.agents.entity_workflow.InferenceClient.text_generation",
             mock_text_generation
         )
 
@@ -277,7 +64,7 @@ class TestParseRecipeEquipment:
             return mock_response
 
         monkeypatch.setattr(
-            "recipe_board.agents.models.InferenceClient.text_generation",
+            "recipe_board.agents.entity_workflow.InferenceClient.text_generation",
             mock_text_generation
         )
 
@@ -298,7 +85,7 @@ class TestParseRecipeEquipment:
             return mock_response
 
         monkeypatch.setattr(
-            "recipe_board.agents.models.InferenceClient.text_generation",
+            "recipe_board.agents.entity_workflow.InferenceClient.text_generation",
             mock_text_generation
         )
 
@@ -316,7 +103,7 @@ class TestParseRecipeEquipment:
             return ""
 
         monkeypatch.setattr(
-            "recipe_board.agents.models.InferenceClient.text_generation",
+            "recipe_board.agents.entity_workflow.InferenceClient.text_generation",
             mock_text_generation
         )
 
@@ -333,7 +120,7 @@ class TestParseRecipeEquipment:
             return None
 
         monkeypatch.setattr(
-            "recipe_board.agents.models.InferenceClient.text_generation",
+            "recipe_board.agents.entity_workflow.InferenceClient.text_generation",
             mock_text_generation
         )
 
@@ -442,8 +229,8 @@ class TestBasicActionsAndDependencies:
         def mock_code_agent(*args, **kwargs):
             return mock_agent
 
-        monkeypatch.setattr("recipe_board.agents.models.CodeAgent", mock_code_agent)
-        monkeypatch.setattr("recipe_board.agents.models.InferenceClientModel", lambda **kwargs: MagicMock())
+        monkeypatch.setattr("recipe_board.agents.parsing_agent.CodeAgent", mock_code_agent)
+        monkeypatch.setattr("recipe_board.agents.parsing_agent.InferenceClientModel", lambda **kwargs: MagicMock())
 
         # Test the function
         result = parse_dependencies(state)
